Fixes #10194. Fixes #10283. Refactor of 708 caption handling.
Note there are a couple minor issues. The default OSDSubFont is not as readable as the old 708 default font and the font can not be reset without restarting mythfrontend. These issues aren't really caused by this changeset though, so I'm applying this as is and leaving those issues to be fixed in a subsequent patch.