Revise Videos storage group fall through for slave backends.
This reworks the strategy used to determine fall through paths for slave
backends to scan. Instead of blindly scanning all directories on all
hosts, this now only scans those directories defined on the master
backend if a slave backend has none of its own.

This also corrects a potential flaw in the previous version where the
iterator could exceed the reach of the list, and perform an invalid
dereference.