Add IPTVStreamHandler.
This is a bit different from stream handlers used for other recorders because the
stream handler is tied to a bind address and port[s] instead of a recording device.

The purpose is the same, to allow overlapping records. But in this case the stream
handler is changed on every channel change.