Fix looking metadata lookup with bad guide data.
When guide data is extremely poor and attempting to perform a
"hail mary" lookup to try to find some valid result, the subtype
of the resulting lookups was not being set to probable television,
so the results were not definitively television.

Now, the results from the generic recording/TV lookup are marked
as such, and thus when the selection is made from the list,
the result is known to be TV and populates the dialogs properly.

Fixes #10343, tested using the same example after having removed
all the other useful guide data to reproduce the issue.