Close unused stdout/stderr on child processes
If UseStdout and UseStderr are not set, after forking, just close the dangling
stdout and stderr file descriptor (as appropriate).  This will keep the output
of the child process from polluting the console of the parent process.  If you
actually want to capture the output, use the UseStdout and UseStderr flags
as needed.

Closes #10143