Fix position/duration display when recordings finish while watched.
The frames reported in the DONE_RECORDING message can be bogus, most
likely due to a race condition or logic error in the recorder.  That
issue will be handled later.  Since the message contains a good time
in seconds value, which is what we want anyway, use it to pass the
finished recording length to the player.