Subtitles: Improve the vertical layout of CEA-708 captions.
The problem is that we were expecting the bounding box height of
caption text to be equal to the font pixel size, which is generally
untrue and can in fact be very different.  As a result, CEA-708
captions could have too much extra space between the lines.  Now, we
target line spacing based on the desired pixel size, taking care not
to let lines overlap.

Refs #10333