Set some sane settings for UDP sockets in IPTVStreamHandler.
This sets a larger than default buffer size so we drop fewer packets.

This also adds support for joining a multicast group. To test change
addr in IPTVChannel::Tune() to a multicast address.

Because setting the receive buffer is not supported by QUdpSocket
and setting joining a multicast group using QUdpSocket requires the
latest Qt, this uses the POSIX API for these things and uses
QUdpSocket::setSocketDescriptor() to assign the socket we've opened
to QUdpSocket for further handling.