Add new 'startyear' and 'endyear' strings to the available recording metadata.
Historically we've never made the year part of recording start/end strings but lots of people have recordings dating back over a year so we need to allow themers to display this information. The simplest fix without breaking existing themes is to expose these new strings that the themer can use in a template with the existing time/date strings.