Metadata Classes: Parse various additional strings from XML.
This has some application in fixing the issue where items at TTVDB
with no seasons or episodes defined, or no 1x01 defined, are not
currently able to get artwork for recordings.

This is only the tip of the iceberg, as there will also be grabber
changes needed to get metadata at the series level, and ultimately
some additional logic to fall back to series level metadata and
artwork when nothing else is available.  It should ultimately
be possible, however, to get artwork for this unusual material.

This also has a greater application to future goals for video library
grouping and UI.