Fix default Network Control port in mythfrontend.
Though the default used by mythfrontend's settings UI is 6546, the code
that starts the NetworkControl was specifying a default of 6545 (which,
incidentally, is the default for the mythlcdserver) in its call to
GetNumSetting().  This typo likely didn't affect anyone because
mythfrontend settings would have populated the default value into the
database, so the default specified by the GetNumSetting() call wouldn't
have been used, except on systems where users have deleted stuff
directly out of the database.

Thanks to Raymond Wagner and David Engel for helping me notice the typo.