Fix recording deletion to always stop recording too.
If AutoExpireInsteadOfDelete was enabled and a recording was still in
progress when it was "deleted", it would keep recording.  There was
code in MainServer::DoHandleDeleteRecording() to stop the recording,
but because of a bad assumption, that code never got executed.  This
change unconditionally stops the recording.  The code for the
AutoExpireInsteadOfDelete disabled case already does this.

The bad assumption was that the recstatus was accurate and would be
either rsRecording or rsTuning for an in-progress recording.  Because
only the chanid and recstartts are passed to the backend for delete
recording requests, the remaining information needed to populate the
recinfo object is pulled from the recorded table and recstatus is
unconditinally set to rsRecorded.

This same bad assumption is made in a few other places.  It's possible
that other "finished recording" processing masks any other problems
caused by it.  The proper fix is probably to pass a complete
ProgramInfo or RecordingInfo object to the backend.  I'll let that
wait until 0.26 or later unless another problem is noticed.