Better handling of exiting playback of a program in use elsewhere.
If a recording contains a cutlist with a final cut region at the end,
and EndOfRecordingExitPrompt is set to true, and the recording is in
use elsewhere (e.g., playing on another frontend) when the end is
reached, it gets into a loop of popping up the "can't delete because
it's already in use" dialog while the final cut region continues
playing in the background.  This is fixed by exiting without waiting
for user input in that situation.