Add hdmi audio device type detection
This commit add ELD (EDID Like Data) detection. It allows myth to detect the various capabilities of the audio device connected via HDMI or DisplayPort.
At present, this information is informative only (except for StereoPCM) and is displayed in the device description in the Audio Settings screen.
It will also show if a device is connected or not on the HDMI or DP port making it much easier to identify which HDMI ALSA device to use (there are often 3 presented, when only one is actually working)

To benefit from this option, you must run alsa drivers >= 1.0.25 or linux kernel 3.2 or later.

StereoPCM advanced user option is now detected automatically if the device ELD is available.
Setting it in the advanded settings will only override this option. As the worse side effect should it be detected improperly is that AC3 re-encoding will occur; I believe the benefits overweigh the downside (if there are any).

This commit includes other minor cosmetic fixes, such as allowing translating the default audio device names in the audio setup screen.

In the future, should ELD be available, the aim is to completely remove the use for audio user settings when using HDMI or DP audio.