MythUI: Fix scrolling with the mouse in MythUIWebBrowser
The scroll position, when triggered by the mouse wheel or by dragging the Qt
scroll bars, was overwritten by the value of the scroll animation even if the
animation wasn't active.  A method IsActive() is added to MythUIAnimation to fix
this.

Scrolling with the mouse wheel when MythUIScrollbars are used is now possible by
adding a wheelEvent() method to MythWebView.