Fix recent regression with metadata lookup.
Some of the behaviors have changed recently (allowing fallback to
series metadata) but that also introduced a regression where 1x01
was being set as the season and episode for items where the season
and episode could not be determined.

Remove the old hack that resulted in this regression... hopefully
this doesn't introduce any other new regressions.