Fix bluray playback when path to bluray mountpoint contains a trailing /
Bluray playback started from either mythfrontend or mythavtest would fail if the path provided contained a trailing / (such as /media/cdrom0/).
Issue lied with ProgramInfo::GetPlaybackURL() trying to find if the file was local. Now do not attempt any file path processing if the URI is for DVD or BD playback

Fixes #10340 (Part 1)