Subtitles: Remove the OSDSubFont setting.
Instead, FreeMono is used for CEA-608, teletext, and text (SRT)
subtitles, as well as the default CEA-708 font.  The issue is that
CEA-608 and teletext layout requires a mono-spaced font, with often
unpleasant results when the previous default FreeSans or another
proportional font is used.  The intention is to make all font choices
themeable in 0.26.