Cutlist editor: Update the edit bar for in-progress recordings.
When editing an in-progress recording, the edit bar only updates when
the cutlist is modified, but not when the recording length changes.
However, the position indicator does get updated, resulting in the
edit bar graphics getting out of sync.

This is fixed by disabling the (seemingly premature) optimization that
updates the edit bar only when the cutlist changes.  This fix will
need to be reevaluated after the 0.25 release.

Refs #10329