Make jump points work during playback when the OSD is present.
This works by sending an EXIT_TO_MENU event at the start of jump point
execution.  Any existing TV playback instance will receive this message
and manually shut down playback, including tearing down the OSD.

This is not the fully general solution desired, but it is a big
improvement until then.

One issue is that an editing session is not shut down cleanly.  That
will be addressed in a subsequent commit.

Refs #7939