Exit cutlist edit mode cleanly when executing a jump point.
When the user executes a jump point in the middle of an editing
session, make sure the edit session closes cleanly.  The work is not
explicitly saved, but the auto-save state is retained in the database
so that it will be auto-loaded next time the user edits that program.

This makes a slight modification to the MythPlayer class and therefore
the binary version is bumped, so be sure to do everything that
entails.

Refs #7939.