Fix LiveTV switching by videosource to honor livetvorder.
Change the GET_FREE_RECORDER_LIST protocol request to return the
cardids in the desired order based on livetvorder.  Change TV to not
re-sort the cardids when mapping sources to an inputid.

N.B. There is still a corner case involving analog cards connected to
multiple sources that could still result in the wrong input being
chosen.  Fixing that would require a bigger change including a
protocol change and can wait until after the 0.25 release.