Fix minor problems editing in-progress recordings.
Revert 5d7c01d08, and add logic that also updates the edit bar when
the total number of frames in the recording changes.

Modify 1ce3d21e3 to use a better test for an in-progress recording.
If the recording finishes during the editing session, the original
test would continue to indicate that the recording was in-progress.

Binary version is updated.  Refs #10329.