Subtitles: improve the layout.
With the existing PAD_WIDTH value, there was still some clipping going
on, for example an italic "W" at the end of the line with the Droid
Sans Mono font.  Simply increasing PAD_WIDTH would have made the extra
spacing between adjacent but differently-formatted text chunks much
more noticeable.

This is improved by increasing PAD_WIDTH but no longer adding padding
between chunks.  A consequence is that the background rectangle now
has the possibility of clipping the text to its left, so there is a
final step that moves all text objects to the front (rather than
drawing background and text in separate passes).

Refs #10194.