Fix the scheduler to not write incorrect oldrecorded entries.
Don't modify the recording status of programs that aren't eligible to
be rescheduled.  Previously, the scheduler properly ignored most
status changes caused by this oversight.  The improved missed
recording detection changed that and allowed some status changes to
get rewritten to oldrecorded.  In the particular case of stopping or
deleting an in-progress recording, the status could change from
rsRecorded and affect future duplicate detection.