Fix the faulty exit code 139 issue in Ubuntu systems
I think it might be an oddity in eglibc, but the status return from waitpid()
is buggered sometimes for signalled children.  Rather than showing up as signal
11 for a segfault, it shows up as an exit value of 139.  This is the most
common offender, but signal 6 (exit value 134) shows up sometimes as well.

The workaround limits the valid exit values of the child to being 0-127.  An
exit value above that is assumed to be a buggered up signal return and is
byteswapped.

As this is most prevalent with mythcommflag, the workaround is only active
for mythcommflag (for now).  If other MythSystem calls need this added, we
can add it later.