Spelling consistency fixes
Replace all (currently) known occurences of translatable strings using the British
and/or Canadian spelling with the US spelling since this is the default spelling
which should be used throughout the application.

Mapping these back to the proper spelling for each variant of English will be done
using the translation files like the rest of the strings.

After applying this the only translation context which could possibly contain words
with a spelling other than the US one will be the "ThemeUI" context were the theme
maintainers will be free to use the spelling they prefer.