MHEG: Fix race condition when restarting the DSMCC after a channel change
When a channel change occurs, the DSMCC file cache must be emptied
so that a consistent filesystem is rebuilt from the new carousel.
MHIContext::Restart, which flushes the cache, is called from TV::ITVRestartv
when the channel or card is changed,  However, the ringbuffer continues
to deliver packets from the old stream for a short period after the
change.  This can produce an invalid file structure which can result
in missing text pages or a failure to start the MHEG app.

This change clears the filesystem cache when a NetBootInfo descriptor
is presented in the PMT, which only occurs after the new stream has started.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>