As noted in SHA1:9551bce2, MainServer doesn't know the actual recording status because it defaults to rsRecorded when the recordinginfo is initialized from the recorded table.  MainServer doesn't account for this and tries to make any needed state change which always ends up staying rsRecorded.  MainServer then passes this bad status to the scheduler.  If the recording had already failed, the rsFailed status gets overwritten.
Since the scheduler knows the correct status, change the status to
rsUnknown in MainServer before updating the scheduler.  The scheduler
now notices the special value and applies any needed state change
itself.  In the case of a failed recording, the status now remains
rsFailed.

Also, remove the duplicate code in MS::DoHandleDeleteRecording() used
to stop an in-progress recording and use MS::DoHandleStopRecording()
instead.  Try to avoid unnecessary reschedules when deleting programs
from the Deleted and LiveTV recording groups.

Fixes #10282