Subtitles: Apply OSDCC708TextZoom to CEA-608 captions as well.
If appropriate care is taken, there is no reason not to apply the text
zoom setting to CEA-608 captions as well as CEA-708 and text captions.
This provides a consistent look across all the caption types.

If the zoom factor would cause the vertical layout to fall outside the
safe area, we try to shift and squeeze the lines back into the safe
area, while also ensuring they don't overlap.