MythZoneMinder: Add support for the deep filesystem hierarchy for events.
ZoneMinder can use 2 different filesystem layouts to store the event images.
This adds support for the layout you get when you turn on USE_DEEP_STORAGE.

This also changes the date/times in the event and event player screens to use
MythDateTimeToString() for consistency with other screens.

NOTE: This bumps the mythzmserver protocol version so both the server and all
clients need to be updated.

Fixes #10151.