Subtitles: Rename the OSDSubFont setting to DefaultSubtitleFont.
This setting is used for all types of captions: CEA-608, CEA-708,
text, and teletext.  The problem is that CEA-608 and teletext are
supposed to use a monospaced font for proper display, but the default
setting was FreeSans, a proportional spaced font.  Most users are
likely to have FreeSans stuck in their database, if they've ever tried
displaying captions/subtitles or visited the Playback OSD settings
page.

Subtitle/caption users may need a one-time revisit of the Playback OSD
settings page if they prefer a font other than FreeMono.