Refs #8941. Make sure we set the channum when switching to a new card with either a specific chanid or inputid.
When switching cards with a chanid or inputid we want to end up
on a specific input, but currently this only happens if channum
is also set (or there is only one input on the card).

Note: I do not have a setup that allows me to test this code
works as intended, but the code should at least not make
anything worse.