Totally rework the AudioReencodeBuffer part in transcode
The fifo mode of transcode got completely messed up with the advent of HLS
mode for mythtranscode.  The core issue was that the video and audio were
being consumed in different ways, and were no longer well synced.  The video
was being queued, and the audio just accumulated until the each video frame
was dealt with.  This caused a majority of the video frames to be dropped
to attempt to resync to the audio, and unnecessarily.

I changed it to actually queue up the audio buffers, and allow the consumer
to pull from that list up until a given timecode, keeping the audio and video
in sync enough for the existing resync code to deal with it.

This has so far been tested in fifo mode only, and with no cutpoints.

Fixes #10199, may affect others as well.

This may break HLS if audio frame sizes differ between input and output, but
this is being worked on, and should be out shortly as well.