Rewrite configure for mythplugins
Previously, configure would only assess if a header or a library were available by looking into a static set of directories. This was very restrive and created multiple issues:
- A library may be present in a particular location, but not be available if say it's been compiled for another architecture
- On a mac, libraries aren't all in one place, in particular OS X Frameworks.

As such, many modules were disabled on a mac build, while they shouldn't have been.

So we use a similar system as used in mythtv's configure script. We actually try to use the library and header to determine if it is available or not.

All the compiler options are retrieved from the mythconfig.mak generated during mythtv's main configure run

Change how we assess the availability of OpenGL. If OpenGL is enabled in the frontend, assume it is available in the plugins