Use QNetworkAddressEntry to get proper broadcast address for udplistener
This restructures the listen address search to store
QNetworkAddressEntry values, rather than simple QHostAddress ones. This
new type provides access to the broadcast address associated with each
IP address, allowing proper selection of broadcasts to listen on for
message popups. This also moves storage and selection of addresses out
of MythCoreContext and into ServerPool, as well as stages a location to
trigger listen server resets, for use when changing IP addresses through
the web configuration utility.