Fix universal package build with new osx-builder
For darwin/OSX, ffmpeg's dynamic libs were built with an id including the full installation path. This meant that upon linking to those libs, mythfrontend would use an absolute path to the libav* ; making the universal build invalid.

Change the linker flags for creating dynamic libraries to remove the absolute path and add headerpad_max_install_names so we can later update the path while creating the application bundle and not get an error that there's no space left.

Conflicts:

	mythtv/configure
[cherry-picked from d1a6cc88]