Fixes #10414. Ensure sufficient buffering and sleep longer when there is no data.
The HDHomeRun StreamHandler depended on a not-well known setting for the buffer size and was using more CPU than it should checking for more data when there was no more data on the wire. This bakes in a reasonable value for the buffer size and sleeps longer after thre is no data.