Add link to disable xslt translation of xsd in browsers (also cleaned up display of array types).
By default, the xsd & wsdl include xml processing instructions
to tell a browser to use a specific xslt file to translate the
schema into html.

I was not able to figure out how to instruct the browser to
ignore this xslt, so I added a query string parameter that
would stop the server from adding the processing instruction.

This parameter is only needed in a browser.  Any schema tools
should be able to use the plan url since it will ignore the
processing instruction.