MythMusic: Add a nodetype statetype to the playlist tree
This changes the playlist tree to show both a statetype that represents the
nodetype eg artists, albums, genres etc as well as an image that shows the
coverart, artist or genre images where appropriate. This is necessary to
remove the hard coded image filenames allowing the themer to use different
names if required.

This also adds icons for the new Compilations and Compilations Artists nodes
and fixes the missing images for compilation artists.