Fix compilation when using 10.5 SDK.
For some reasons, loading the OpenGL headers fail when using the 10.5 SDK. This is due to some data types not being defined.

So add the framework where those types are defined.

I couldn't find an elegant way to query what the current SDK is, so instead if compiling under 10.5 SDK, you define an extra -D_USING_105SDK=1 to CFLAGS and CXXFLAGS