Fix new piano visualiser.
STL std::vector doesn't have a data() class, so use a pointer to the first element instead.

I don't see why we wouldn't use QVector to start with, and why use a combination of STL vector and QTL QVector.
But as I want to keep the changes to a minimum during feature freeze, apply the least invasive change

Reference: #10214