Remove limit from 1256 DB update.
The limit, coupled with a post-0.24 DB upgrade, could cause a key
binding collision on exactly one host (and not more than one host) in
the MythTV system when upgrading from a 0.23 schema version to post 0.24
schema version.  The collision could only occur in a multi-host MythTV
system, could not occur on the host performing the DB upgrade(s), and
could only occur on a host that had been used as a frontend with 0.23
but never ran either mythfrontend or mythtv-setup from 0.24.

More specifically, you would have had to have a schema version from
before May 10, 2010 on master (such as a 0.23 schema version) and
upgraded to a schema version from after July 21, 2011 on master (not in
any released version, but will be in 0.25) without ever running a
version of mythtv from between those 2 dates.  Even then, the collision
is unlikely to occur because of how the data is returned.

For users running master, any affected host's mythfrontend or
mythtv-setup will show the warning:

2012-03-06 09:08:55.862679 W  Key Esc is bound to multiple actions in
context Main Menu.

The collision can be fixed by manually removing or changing the binding
on the action you don't want in mythfrontend's Utilities/Setup|Edit Keys
in the Main Menu context on the affected host.

For users running 0.24 versions, there is no collision, and simply
running mythfrontend (or mythtv-setup, but no need to run both) from
0.24-fixes on your frontend hosts will prevent one from occurring when
you upgrade to 0.25, so you have nothing to worry about/nothing to fix.

Refs #10403.
(cherry picked from commit e022c2c9769f53a78355d1bb5950bf33593cfbad)