Change SignalMonitor shutdown so it works for IPTVSignalMonitor
The problem was mainly that the MThread::exit(0); wasn't being called when shutdown from the SignalMonitor dtor because it was calling it's version of SetRunningDesired() not the IPTVSignalMonitor version. But there also seems to be a problem wherein the MThread::exit(0); doesn't register with the QThread, so we now call it every 100 ms until it takes.