Set up IPTVStreamHandler only when we have a StreamData.
With other recorders the StreamHandler lives as long as the device does, but for IPTV we
need to create a new IPTVStreamHandler for every channel we tune to. So instead of creating
one whenever Tune is called we now only create it when the MPEGStreamData is set.

This means we need to unset the MPEGStreamData to pause the IPTVRecorder and set it
when unpausing the IPTVRecorder so RecorderBase::PauseAndWait() has been overridden
to do that.