Add more detail to syslog logging
As a fair number of users are likely to be using syslog rather than file
logging, it has been requested by several people that the thread/file/line
type of detail be added to the syslog output as well.  Since this helps a lot
for debugging issues (when combined with the version string which identifies
the git checkout), I've added it now, as I'd forgotten several times in the
past.