Correct handling of old --refresh-day option.
This converts the --refresh-day input from an unsigned integer to a
stringlist, to allow the option to be used multiple times and accumulate
multiple days as was its previous behavior.  This option is still marked
as deprecated, and will be removed in favor of the new --refresh option
in a later version.