Make XML plist serializer dates work with Apple plist parser.
Apple is apparently picky about the <date> explicitly stating that it is Zulu time
and having the Z at the end.  Apparently in most versions of Qt, Qt::ISODate doesn't
include it, and it may be added in 4.8... the only way to avoid the versioning
mess is to just explicitly specify the date format in the serializer.