Subtitles: Improve layout of zoomed CEA-608 captions.
Thanks to Gregorio for the patch that improves the CEA-608 layout, in
particular when the zoom factor is less than 100%, by horizontally
centering the caption region and vertically snapping lines to the top
or bottom of the safe area.  One key difference from the original
patch is that horizontal centering is disabled when then zoom factor
is 100% or more.  While horizontal centering might be more visually
appropriate, I'm not comfortable that it becomes markedly different
from the "equivalent" CEA-708 layout.

Refs #10402.

Signed-off-by: Jim Stichnoth <jstichnoth@mythtv.org>