Fix RingBuffer starvation when playing .nuv files.
In the original code, the estimated bitrate for .nuv files was set to
0, and downstream code would figure out a suitable blocksize for
ringbuffer fills.  After the low-bitrate optimizations in #9824, the
blocksize ended up being reduced to 2KB.  This may be sufficient for
local files, but often causes ringbuffer starvation and stuttering
playback for remote files.

Fixes #10212.  If similar behavior is seen post-commit, please file a
new ticket.