XML plist serializer: Fix dictionary in array serialization.
All items need keys, *except* for items in arrays, where the
expectation is that each item in the array is one of the parent
key.

Previously, each dictionary in an array was being parsed as two items--
one invalid (the key) and one valid (the value) leading to some interesting
crashes.