Partially revert 5dccbd5f6b3019a67ecc69213ee405bba4b30994
While this worked when using Qt frameworks on a mac, when using Qt libs I was seeing the same issue as in edae94e0458a9b7fc054adb2f0a1df7882929775.
Basically, it would like with old mythtv libs.

Rework the changes so we do reset QMAKE_LIBDIR_QT when not using Qt frameworks on a mac