Subtitles: Try to reenable the previously selected caption type.
During DVD playback with AVSubs enabled, once in a while the decoder
disables and reenables subtitles.  If the DVD also contains cc608
captions, the caption priority order would favor cc608 and the user
would need to reselect AVSubs every time this happens.  To fix this,
we keep track of the previously selected caption type and default to
that if possible when captions are turned on.  This is in addition to
the existing code that tries to reselect the previous AVSubs track
when AVSubs are enabled.