Performance/bug fix in media monitor ScanMediaType().
Symlinks to other places (like /Applications on OS X disk images) can cause
unwanted scanning of irrelevant directories and a big performance hit.
Even worse is symlinks which point to themselves, or a parent directory,
which recurse forever and blow the stack. If anyone really thinks that symlinks
should be used in file counts, we can revisit and add recursion limit rules?