Fix Theme Chooser downloads on remote frontends for 0.25
Some users are reporting issues with downloading themes on remote
frontends.  This patch modifies the Theme Chooser to download the
theme package directly on the frontend rather than having the
master backend download the theme first then having the frontend
download from the master backend.  This is a temporary fix to get
the Theme Chooser working for the most number of users for 0.25.
After release, this fix may be reverted at some point if the cause
of the reported issues can be found and fixed.