Fixes #10054. Fixes handling of segmented SDT and PAT tables.
The code was attempting to verify if it had seen an SDT for
a given TSID before, and it inadvertantly created an empty
entry rather than just verifying it was missing.

This leads to further errors when the scanner thinks it has
an SDT for a transport when it is in fact missing it.

The same logic issue was found in the PAT handling.

Based on original patch by
Argus <pthorn-mythd at styx2002.no-ip.org>