Generalize caching and allow post configuration.
This moves the caching code into a private 'engine' and adds a function
to allow selection and manipulation of that.  Currently, only the null
and file engines are available.  The file engine is the previous
behavior, and the null engine disables caching all together.

from request import set_cache
set_cache(filename='some_file')
set_cache(engine='null')

The filename will be used directly if the first character is '/',
indicating an absolute path. '/tmp' will be prepended otherwise.