Refs #10490. At least a partial LiveTV startup fix.
When we call FinishedRecording() early it appears to cause an extra empty
file to be generated at least some of the time. I'm not exactly sure how
this happens, but calling FinishedRecording before the recording is finished
is actually a bug itself. This fixes this in the case where I was able to

reproduce the problem Peter Townsend experienced. I also tried some other
LiveTV actions without seeing duplicate FinishedRecordings calls, but I
may have missed some code paths.