Allow locale fallthrough for movie images and alternate titles.
When fallthrough is enabled, queries for backdrops, posters, and
alternate titles will not filter for the selected locale, but will
instead sort the results to place those matching the specified locale at
the top of the list.

Also resolves an issue where Request() was incorrectly passing 'None' as
an argument when no locale was set.