Fixes #9830. Refs #10519. Fixes channel change on PVR350.
The drivers for some V4L devices require you to explicitly stop and restart streaming when sending certain ioctl's to the device. We already pause and unpause the recorder when issuing these commands so this adds a close(readfd) to the code that is run on pause and reopens readfd on the unpause code. This also means the DeviceReadBuffer needs to be reset with the new readfd, so we do that as well.