DVR Service: Fix recording deletion API.
As with a few other APIs, this was developed on a slave backend,
where a remote delete request could work.  As the best/correct use
of the APIs is to speak to the MBE, allow the API to work
by sending the deletion mythmessage.

Testing and working with proper UTC input.  Fixes #10540.