Fix popup windows from disappearing in MythCenter-wide when using the virtual keyboard.
Adapted from a patch by Harley Peters <harley@thepetersclan.com> for
the TintedGlass theme.

Disable some "AboutToHide" animations in the MythCenter-wide theme for
popup windows that can access the virtual keyboard.  The animations
cause the window to disappear when the virtual keyboard is displayed
and consequently don't show what is typed.  The "AboutToShow"
animations remain in effect, and unfortunately, cause some flickering
when the virtual keyboard is dismissed.  Even with the flickering,
this is better than not being able to see what is typed.  Hopefully a
better fix can be made to 0.25-fixes later.